More cleanup
Separate violations code, simplify modules
